Calorie King/requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic>=0.40.0
pillow>=10.4.0
flask-login>=0.6.3
gunicorn>=21.2.0
